fix: out of range warning component moved into parent folder
